refactor(check): descend the thinking judgment into nika-check
The judgment landed in the CLI's models rung and pushed nika-cli 60
prod lines over the 15000 crate wall. The judge is pure (workflow in,
findings out), so it descends to the check crate beside the hints layer
that already reads the same catalog reasoning fact — the brief's second
offered home — as thinking.rs with its ThinkingFinding rows. The CLI
keeps a thin wrapper mapping rows into the rung's ModelFinding, so the
three fold sites (check verdict, slot-only gate, dry-run swap) and the
wiring pin are untouched. The judgment unit tests moved with the code;
the boundary mutation was re-proven in the new home (budget at the cap
goes red when the compare slips past it). A side effect worth keeping:
the machine surfaces (MCP) can now reach the judgment without
re-deriving it.
