+++ crates/nika-cli/src/verbs/check/models_rung.rs
@@ -213,119 +213,20 @@ pub(crate) fn unresolvable_models(
     audit
 }
 
-/// The `infer.thinking` judgments the parse cannot make: the parser
-/// validates each field's TYPE only — the two laws below are cross-field
-/// (budget vs cap) and cross-fact (the seat's reasoning capability), and
-/// the dispatch copies `budget_tokens` verbatim into the provider call —
-/// nothing downstream re-judges either. Both were false greens: ✔ check,
-/// then the refusal (or the silently dropped budget) at run.
-///
-/// - **the budget lives UNDER the cap** — with `budget_tokens` and
-///   `max_tokens` both declared, the budget must be smaller: the
-///   reasoning share lives inside `max_tokens`, so a budget ≥ the cap
-///   leaves no room for the visible answer and the provider refuses the
-///   call at run.
-/// - **the seat can reason** — `thinking: { enabled: true }` on a model
-///   the catalog KNOWS cannot reason is a dead declaration (the wire
-///   400s or drops the budget). Refused only when the catalog carries
-///   the model row — a model the catalog does not know is the
-///   `catalog_warning` lane's advisory, never this refusal (the
-///   snapshot is dated; silence beats a wrong refusal).
-///
-/// A templated seat (`${{ }}`) is judged THROUGH its declared default —
-/// the same `static_literal_of` law as [`unresolvable_models`]; one with
-/// no literal default defers to the run, making no claim.
+/// The `infer.thinking` judgments, folded into this rung's findings.
+/// The judge is [`nika_check::thinking_findings`] (nika-check · the laws
+/// and their scope live with it): the parse validates each field's TYPE
+/// only, so the cross-field (budget vs cap) and cross-fact (the seat's
+/// reasoning capability) laws descend to the check crate — a judgment
+/// the MCP/machine surfaces can reach without re-deriving it here. The
+/// lane lived in this file until the 15k crate wall moved it
+/// (2026-08-25); the fold sites (`check` verdict · slot-only gate ·
+/// dry-run swap) are unchanged.
 pub(crate) fn thinking_findings(wf: &nika_schema::raw::RawWorkflow) -> Vec<ModelFinding> {
-    let mut findings = Vec::new();
-    for task in &wf.tasks {
-        let nika_schema::raw::RawAction::Infer(action) = &task.value.action else {
-            continue;
-        };
-        let Some(thinking) = &action.thinking else {
-            continue;
-        };
-        if !thinking.value.enabled {
-            continue;
-        }
-        let id = task.value.id.value.as_str();
-        // The budget-vs-cap compare is seat-independent: both fields are
-        // literal u32s by the time the parse admits them.
-        if let (Some(budget), Some(max)) =
-            (thinking.value.budget_tokens, action.max_tokens.as_ref())
-            && budget >= max.value
-        {
-            findings.push(ModelFinding::new(
-                seat_label(wf, action),
-                vec![id.to_owned()],
-                format!(
-                    "`thinking.budget_tokens` ({budget}) on `{id}` must stay UNDER \
-                     `max_tokens` ({}) — the reasoning share lives inside the cap, so a \
-                     budget ≥ the cap leaves no room for the answer and the provider \
-                     refuses the call at run",
-                    max.value
-                ),
-            ));
-        }
-        // The reasoning-seat law needs a statically known seat.
-        let Some(seat) = action.model.as_ref().or(wf.model.as_ref()) else {
-            continue;
-        };
-        let judged = if seat.value.contains("${{") {
-            let Some(literal) = nika_check::static_literal_of(wf, &seat.value)
-                .and_then(|v| v.as_str().map(str::to_owned))
-            else {
-                continue; // no literal default — the run judges
-            };
-            literal
-        } else {
-            seat.value.clone()
-        };
-        let Some((provider, name)) = judged.split_once('/') else {
-            continue; // no provider prefix — the resolver's refusal owns that class
-        };
-        // The refusal needs the catalog to POSITIVELY know this exact
-        // seat — the two lanes the run itself trusts (the
-        // `catalog_warning` precedent): a provider row (the binary's own
-        // nicknames + wire ids) or an EXACT pricing pattern. A fuzzy
-        // pricing match never counts: it exists to absorb dated variants
-        // for a cost estimate, not to license a refusal — and a model
-        // the catalog has never heard of keeps the defaults
-        // (`reasoning = false` = "no evidence it reasons", NOT "it
-        // cannot"), so without this gate every newly shipped model would
-        // refuse a legal `thinking:`.
-        let known = nika_catalog::find_provider(provider)
-            .is_some_and(|row| row.models.iter().any(|m| m.id == name || m.model == name))
-            || nika_catalog::all_pricing().iter().any(|p| {
-                p.provider.eq_ignore_ascii_case(provider)
-                    && (p.model_pattern == name || p.model_pattern == judged)
-            });
-        if known && !nika_catalog::model_capabilities(provider, name).reasoning {
-            findings.push(ModelFinding::new(
-                judged.clone(),
-                vec![id.to_owned()],
-                format!(
-                    "`thinking: {{ enabled: true }}` on `{id}` seats `{judged}` — the \
-                     catalog knows this model cannot reason, so the declaration is dead \
-                     at the wire (the budget is dropped or the call refused); remove \
-                     `thinking:` or seat a reasoning-capable model"
-                ),
-            ));
-        }
-    }
-    findings
-}
-
-/// The seat label a thinking finding prints — the task's `model:` or the
-/// envelope default, `-` when neither reaches the task.
-fn seat_label(
-    wf: &nika_schema::raw::RawWorkflow,
-    action: &nika_schema::raw::RawInferAction,
-) -> String {
-    action
-        .model
-        .as_ref()
-        .or(wf.model.as_ref())
-        .map_or_else(|| "-".to_owned(), |s| s.value.clone())
+    nika_check::thinking_findings(wf)
+        .into_iter()
+        .map(|f| ModelFinding::new(f.model, vec![f.task], f.why))
+        .collect()
 }
 
 /// The rates the preflight shows BEFORE the first run: each model the
@@ -379,14 +280,8 @@ pub(crate) fn pricing_section(
 
 #[cfg(test)]
 mod tests {
-    use super::*;
     use nika_schema::parser::{ParseMode, parse};
     use nika_schema::source::FileId;
-
-    fn findings_of(yaml: &str) -> Vec<ModelFinding> {
-        let wf = parse(yaml, FileId::new(0), ParseMode::Strict).expect("fixture parses");
-        thinking_findings(&wf)
-    }
 
     fn infer_wf(model: &str, max_tokens: &str, thinking: &str) -> String {
         format!(
@@ -395,130 +290,12 @@ mod tests {
         )
     }
 
-    /// The budget-vs-cap law: the parse typed both fields as u32 and
-    /// stopped there; the dispatch copies the budget verbatim, so a
-    /// budget ≥ the cap audited green and the provider refused the call
-    /// at run. The finding fires at check, before a token is spent.
-    #[test]
-    fn a_budget_at_or_over_the_cap_is_a_finding() {
-        for (budget, max) in [("100", "100"), ("101", "100")] {
-            let f = findings_of(&infer_wf(
-                "mock/echo",
-                max,
-                &format!("{{ enabled: true, budget_tokens: {budget} }}"),
-            ));
-            assert_eq!(f.len(), 1, "budget {budget} vs cap {max}: {f:?}");
-            assert_eq!(f[0].tasks, vec!["t"], "{f:?}");
-            assert!(
-                f[0].why.contains("budget_tokens") && f[0].why.contains("max_tokens"),
-                "the finding names both fields: {}",
-                f[0].why
-            );
-        }
-
-        // Control pair: a budget UNDER the cap makes no finding (a
-        // compare that always fires is the same defect, mirrored).
-        let ok = findings_of(&infer_wf(
-            "mock/echo",
-            "100",
-            "{ enabled: true, budget_tokens: 50 }",
-        ));
-        assert!(ok.is_empty(), "the legal twin stays clean: {ok:?}");
-    }
-
-    /// The compare needs BOTH literals: a budget without a declared cap
-    /// (the provider default governs) and a cap without a budget (the
-    /// provider's own thinking bound governs) make no claim.
-    #[test]
-    fn the_compare_needs_both_sides() {
-        let no_cap = findings_of(
-            "nika: w\ntasks:\n  t:\n    infer:\n      prompt: hi\n      model: mock/echo\n      \
-             thinking: { enabled: true, budget_tokens: 5000 }\n",
-        );
-        assert!(no_cap.is_empty(), "no declared cap, no compare: {no_cap:?}");
-        let no_budget = findings_of(&infer_wf("mock/echo", "100", "{ enabled: true }"));
-        assert!(
-            no_budget.is_empty(),
-            "no declared budget, no compare: {no_budget:?}"
-        );
-        let disabled = findings_of(&infer_wf(
-            "mock/echo",
-            "100",
-            "{ enabled: false, budget_tokens: 100 }",
-        ));
-        assert!(
-            disabled.is_empty(),
-            "enabled: false declares no thinking: {disabled:?}"
-        );
-    }
-
-    /// The reasoning-seat law: `enabled: true` on a model the catalog
-    /// KNOWS cannot reason (gpt-4o-mini is a catalog row with
-    /// `reasoning = false`) is a dead declaration — refused at check.
-    #[test]
-    fn thinking_on_a_known_non_reasoning_seat_is_a_finding() {
-        let f = findings_of(&infer_wf(
-            "\"openai/gpt-4o-mini\"",
-            "500",
-            "{ enabled: true, budget_tokens: 100 }",
-        ));
-        assert_eq!(f.len(), 1, "a dead declaration is a finding: {f:?}");
-        assert_eq!(f[0].model, "openai/gpt-4o-mini", "{f:?}");
-        assert!(f[0].why.contains("reason"), "{}", f[0].why);
-
-        // Controls — every neighbor the refusal must NOT eat:
-        // a reasoning seat (o3) · a seat the catalog does not carry
-        // (the catalog_warning lane's advisory, never this refusal) · a
-        // templated seat with no literal default (the run judges it).
-        for yaml in [
-            infer_wf(
-                "\"openai/o3\"",
-                "500",
-                "{ enabled: true, budget_tokens: 100 }",
-            ),
-            infer_wf(
-                "\"mock/echo\"",
-                "500",
-                "{ enabled: true, budget_tokens: 100 }",
-            ),
-            "nika: w\ninputs:\n  seat: { type: string, required: true }\ntasks:\n  t:\n    \
-             infer:\n      prompt: hi\n      model: \"${{ inputs.seat }}\"\n      \
-             max_tokens: 500\n      thinking: { enabled: true, budget_tokens: 100 }\n"
-                .to_owned(),
-        ] {
-            let f = findings_of(&yaml);
-            assert!(f.is_empty(), "no refusal here: {f:?}\n{yaml}");
-        }
-    }
-
-    /// A templated seat is judged THROUGH its declared default — the
-    /// rung's via-default law applied to the reasoning fact: a
-    /// `${{ const.seat }}` whose default cannot reason is the same dead
-    /// declaration, named as the default that was judged.
-    #[test]
-    fn a_templated_seat_is_judged_through_its_declared_default() {
-        let f = findings_of(
-            "nika: w\nconst:\n  seat: \"openai/gpt-4o-mini\"\ntasks:\n  t:\n    infer:\n      \
-             prompt: hi\n      model: \"${{ const.seat }}\"\n      max_tokens: 500\n      \
-             thinking: { enabled: true, budget_tokens: 100 }\n",
-        );
-        assert_eq!(f.len(), 1, "the declared default is judged: {f:?}");
-        assert_eq!(f[0].model, "openai/gpt-4o-mini", "{f:?}");
-    }
-
-    /// `agent:` carries no `thinking:` field and `exec`/`invoke` no seat
-    /// law — the judgment is infer-scoped by construction.
-    #[test]
-    fn non_infer_tasks_make_no_thinking_claim() {
-        let f = findings_of("nika: w\ntasks:\n  t:\n    exec: { command: [\"echo\", \"ok\"] }\n");
-        assert!(f.is_empty(), "{f:?}");
-    }
-
     /// The wiring pin: the judgment must reach the VERDICT — a finding
     /// computed but never folded into `clean` is the false-green class
     /// this arc exists to close. Drives the real `check` verb end to
     /// end; deleting the `findings.extend(thinking_findings(..))` fold
-    /// turns this red while every judgment test above stays green.
+    /// turns this red while the judgment's own tests (nika-check's
+    /// `thinking` module) stay green.
     #[test]
     fn a_thinking_finding_turns_the_check_red() {
         let dir =
@@ -554,5 +331,23 @@ mod tests {
         let ok =
             crate::verbs::check::run(ok_path.to_str().expect("utf8"), false, false, None, theme);
         assert_eq!(ok.code, 0, "the legal twin stays green: {}", ok.text);
+    }
+
+    /// The wrapper maps the check crate's rows into this rung's finding
+    /// shape — the model seat, the ONE task, the why, and never a
+    /// conjured spec code.
+    #[test]
+    fn the_wrapper_maps_thinking_findings_into_model_findings() {
+        let wf = parse(
+            infer_wf("mock/echo", "100", "{ enabled: true, budget_tokens: 100 }").as_str(),
+            FileId::new(0),
+            ParseMode::Strict,
+        )
+        .expect("fixture parses");
+        let rows = super::thinking_findings(&wf);
+        assert_eq!(rows.len(), 1, "{rows:?}");
+        assert_eq!(rows[0].tasks, vec!["t"], "{rows:?}");
+        assert!(rows[0].why.contains("budget_tokens"), "{rows:?}");
+        assert!(rows[0].code.is_none(), "engine-local, no conjured code");
     }
 }
